backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
asyncpg==0.24.0
python-dotenv==0.19.0
passlib==1.7.4
python-jose==3.3.0
pydantic==1.8.2